Refactorisation SQS : nommage, config, stratégie de réessai, ordre des méthodes
