+++ src/main/java/fplarache/smartlib/notifications/NotificationServiceApplication.java
@@ -1,4 +1,4 @@
-package com.example.notificationservice;
+package fplarache.smartlib.notifications;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
