+++ src/main/resources/application.properties
@@ -12,3 +12,6 @@ spring.mail.properties.mail.smtp.starttls.enable=true
 spring.mail.properties.mail.smtp.starttls.required=true
 
 aws.region=eu-north-1
+sqs.queue.url=https://sqs.eu-north-1.amazonaws.com/774305596814/notification-sqs
+spring.profiles.active=dev
+
